Extract Runtime.BranchMerge from workflow runtime
Move the pure branch-state merge (reducer / channel / last-write-wins) out of the recursive interpreter into a dedicated module. Structure-only: same behavior, two call sites delegate.
